Lint check for valid class references
Lint already checks that classes registered in the manifest (such as
activities and services) exist and are instantiatable.

This lint check expands this check to also make sure that custom views
and fragments referenced in layouts also exist, as well as class names
registered in analytics config files (requested in issue 41567).

It also fixes the typography detector to not flag IDs as needing
em dashes.

Change-Id: Iee360912f06ac9d42572b8979e28fafbfaef561b