Separate detector metadata from issue
This changeset makes a change to how lint maps issues to detectors.

Until now, each issue would name a detector class, as well as the
issue scope required by that detector, to analyze this issue (as well
as incremental analysis scopes). If an IDE wants to replace certain
detectors (for example, to more directly use extra data available to
make the rule more widely applicable or accurate, such as type
resolution) this meant replacing the issue instance itself, which
doesn't work well when other code references an issue by
instance. It's also hacky. Now, with these concerns separate, a tool
can replace the implementation instance of an issue while continuing
to report the same issue.

Since I was making an incompatible change to the Issue class, I took
the opportunity to make several other changes I've been meaning to get
to:

- An issue can now specify more than a single "More Info" URL

- There's a new "Brief" description, typically 4-8 words, which
  captures the essence of the bug. This is used in headers such as the
  Preferences dialog etc, in the HTML report summary instead of ids,
  etc.

- There's a new OutputFormat enum. Until now, the explanation
  attribute of an issue had some markup (* for bold, ` for code, etc),
  and there were multiple methods to get the explanation:

   * getExplanation() -- returned the raw markup text
   * getExplanationAsSimpleText() -- generates plain text version
   * getExplanationAsHtml() -- generates HTML markup (using <b> for
      bold etc)

  This was specific to the explanation, and in particular the
  description could not contain markup for code identifiers.

  Instead, there's now a single getExplanation(OutputFormat) method,
  and similarly getDescription now takes an OutputFormat, as does the
  new getBriefDescription() method. This also allowed many issues to
  start using code markers and bold markers in the issue summary.  The
  HTML report for example takes advantage of this.

Change-Id: I3571b8e104695db3e3d994077bd0bcb492d474d2