Add support for invoking methods on views.
This CL adds support for invoking methods on a Android View
object. To invoke a view, the method name along with all its
arguments are passed to the device via the DDM channel.

The existing commands to layout a view and invalidate a view
are now replaced by just calling the generic invoke view method
operation with appropriate parameters.

Change-Id: I1fe345f9c6d7db0569656256feb091c6f1b664b6