Local sdk.
See java doc of LocalSdk for details.

Background: the goal is to write a new class, LocalSdk,
that will eventually replace the ill-named SdkManager
class. This moves lots of existing SDK parsing code
(scaterred in a few places) into the new LocalSdk.
It does not however change anything to the way the
SDK files are being parsed nor does it change the
existing AndroidTarget objects, in order to make sure
to not introduce behavior changes.

Change-Id: Id1e2d6cf8920c63653337b4b178c940ac496bf9b