SDK Manager: deal with archives that lack a root dir.
All archives unzipped by the SDK Manager should have a
root directory. This was done for convenience and historical
reasons.

The unzip code was written so that it would just drop any
zip entry that isn't located in a top-folder. This was done
as a simplification to drop the top-level folder entry
(which appears a "rootdir/", f.ex.). An unfortunate consequence
is that this drop every leaf files from a zip improperly packaged.

The change is that if a zip entry doesn't have a /, it's a file
located at the root of the zip file hierarchy and we still accept
it and place in the top-level folder. However there's only supposed
to be a single top-level in a zip archive and their name doesn't matter.
If a zip is malformed and has multiple top-level folders, all their
content will be merged together.

Change-Id: I46634583432363aa73e7b006b45dc0c35953f780