Lint: Handle codenames in minSdkVersion, and add 4.3 info
This changeset fixes lint to handle code names in the minSdkVersion
declaration in the manifest (for preview platforms). Prior to this, if
it encountered a preview platform it didn't know about, it would take
the minSdkVersion to be 1, and thereby flag a lot of calls as not
available.

Also, update the SdkVersionInfo class with information about API 18
now that Android 4.3 has shipped.

Change-Id: If39edd3afb534a99c054b0bb2e121c79a30d2ad9