Maintain stats per thread, and other cleanup.
The method info structure maintains per method information,
accumulated across all invocations of that method.

A method could however be called from multiple threads, so
this CL fixes the MethodInfo structure to maintain per thread
statistics.

In addition, this CL introduces a ClockType (thread and global)
to allow changing which clock is used for display.

Change-Id: I672ee41bf516571a8e30503db9c8e56aa562b078