Maintain stats split by caller/callee.
This CL moves all the per method stats into a separate class
MethodProfileData. In addition to the inclusive and exclusive times
for the method itself, this class also stores the inclusive and
exclusive times accounted by each caller of the method and callee
from the method, and invocation counts (cumulative as well as split
by caller and callee).

Change-Id: Ia29fc2a265979f5051bababbe1d57b3262cbd240