Sdklib: Add repository pkg descriptor classes.
Use enums for LocalSdk package types.

Refactor PkgDesc to make actually classes internal and only
expose an IPkgDesc interface. PkgDesc acts as a factory.

Package descriptors are property-containers that contain the
attributes of LocalPkgInfo instances. They will also be used
to contain the attributes of the RemotePkgInfo instances in a
next CL.

Added and updates unit tests accordingly.

Change-Id: I3fb9ab3fd95df728bb17509eb40e19b5ccb4332b