Add issue explanations to lint text output
This CL adds a new option, explainIssues (true by default in the Gradle plugin)
which causes the plain text output of lint (which for example is used by the
lintVital task) to include full issue explanations in the error output.

This helps solve the problem that the one-line error message is often not
sufficient to explain a new issue to a developer, and it may not have been
obvious where to look for more info.

The output can be turned off with

android {
    lintOptions {
        explainIssues false
    }
}

Change-Id: I63190e893dfe366ed053d8705a94d7b73014670a