Switch default minSdkVersion from 7 to 8 in new projects/modules
8 was the default we used in ADT. I think we accidentally started
using Studio because in Studio we obey the actual value supplied
in the template (7) rather than having a hardcoded default (8)
in ADT where the wizard didn't look at the project template
metadata.

One reason we should do this is that several of our templates
require API 8. Another is that we don't even list API 7 on
http://developer.android.com/about/dashboards/index.html
anymore; 8 is the oldest listed so presumably API 7 isn't
relevant anymore.

Change-Id: I778f3525643dfce108eab672386b7f0787560c66