Fix test.
The tests don't actually need to check that these folders
exist. These are only created when the project is built and so they
can fail since these tests don't build them.

What matter is that the model is correct and this is tested separately.

Building the projects in the tests would make the test even slower,
and there are other tests already building them to make sure the
projects keep building.

Change-Id: I8045b7a50d049df77b16d5e33955332c4481b90c