Fix tests and incremental support in manifest.
device test were broken due to unique package name in libraries
fix. The library samples where only partially updated which allowed
building them but not building or running their tests. This CL
fixes them.

Doing this I also encountered an issue in the incremental detection
of the manifest tasks so this fixes that too. The issue is that the
@Nested annotation only work on the direct fields of a class
and doesn't work if the class is a collection (ie it doesn't nest
into the content of the collection).

Change-Id: I6f624ece39717e6581897f014ef77d3cd9ba7cde