Added validation step for selector to ensure their correctness.
We now load all xml files upfront  to build a list of acceptable
tools:selector values. This will be used when a tools:selector is
found on a node to ensure its value is correct.

Found an issue while debugging, we were cleaning node with
tools:selector rather then just removing the attribute, this was
deleting the element the selector was intending to keep !

Change-Id: I17d578c0190708de08894191e11c34bb1c72d4e1