Include platform-tools version in API cache database name
The API database binary cache, used by lint's API checker, currently
includes a version number (used when we change the schema). It also
looks at the modification timestamp of the database file
(api-versions.xml), and if newer than the cache, it regenerates the
binary cache.

However, if you for some reason reinstall an earlier version of the
tools, such that the modification timestamp of that older platform is
more recent than the cache, then the cache will downgrade to that
older version.

To make this work more reliably, we now look up the platform-tools
version string, and include that in the cache file name. This means
that the cache will be unique to each version of the platform tools,
which is what we want since the data is distributed with the platform
tools.

Change-Id: Ie58474374957c81e5c2ff62bdda7d25dc8cf931a