Fix lint fragment detector to recognize implicit default constructors
This check used to be based on analyzing the bytecode, where the
compiler always inserts the implicit constructor such that it appears
as a declared method. In the AST however, it doesn't appear, so we
have to consider the absence of any constructors as the presence of
a default constructor.

Change-Id: If755a1654e9794251fb60d01292e7c81dd017d73