Fix lint gradle check message, and unify targetSdkVersion lookup
First, the lint error message in Gradle files when using a support
library from a revision older than the targetSdkVersion was printing
out the wrong version (compileSdkVersion instead of targetSdkVersion)
and it refered to "the" support library (when there are multiple such
libraries) which I changed to "this" support library instead.

Also unify the way the manifest and gradle lookups for the latest
targetSdkVersion works. Don't complain if not matching a preview
platform.

Change-Id: I678043adb110c07053b754bf3f98c86f1602d198