Made placeholders field common to build types and build flavors.
Up to now, users could only specify placeholders values in the build flavors
part of the build.gradle.
With this change, placeholders values (with the same syntax) can also be
specified in the build type declaration. The build type value will always
override build flavor values for the same key.

Change-Id: I35016a520666bd5f0ae487ffa81cf9a12711c069