Incremental handling of merged items during res merge.
Custom handling for declare-styleable introduced the
concept of merged items during a res merge.

These items are created dynamically during the merge and
are not tied to a DataSet. This means they must be
stored and handled separately in the Merger to support
incremental merges.

This changeset does this and adds a set of tests representing
various scenarios.

It also introduces changes to the merge code in order to properly
support incremental support of merged items.

Change-Id: Ib2b2ae15c3871fec0edc86ae2a2f95bab5405d58