74988: Import process breaks project by depending on latest support library
This CL makes the project import insert *specific* versions of the
support library dependencies, compatible with the compileSdkVersion of
the imported project. It also reorganizes some code such that the code
to look up SDK maven repositories can be shared between the Gradle lint
check, module import, and some similar code in the Android Studio plugin.

Change-Id: Ibf0f11b6d11fc57b91ff3cd7db0e491e8a27abe3