System property fetcher: Misc fixes
This CL adds tests for a couple of scenarios where the property
fetcher was not behaving correctly. One of these was already fixed
in a previous CL, so this CL includes a fix for the second scenario
where the property cache state was not set properly if the response
from the device was empty.

This CL also removes the backoff mechanism where properties weren't
being fetched for a certain amount of time from a previous fetch.
This backoff had the unintended effect of making things look slow
in the IDE because fetching the properties always involved this
backoff delay in case the first attempt to fetch properties failed,
which seems to happen often.

Change-Id: I86fd551633917e8647c2f6102c88747375648bcc