Change java compile task to generic compile task in model
The previous model included a task to allow compiling
only up to the Java code for a given artifact. This did
not include native code, so the new model has a generic
compile task instead.

The old java compile task is still there for now but will be
removed in 1.0

Change-Id: I030a069130613fee5692123966d1acc1d0926996