75879: Don't force full translation for regional strings
If your base strings in the values/ folder are in English,
and you customize a few of these in (for example) values-en-rGB,
lint currently flags these strings as missing from the values-en
locale, so you'd need to go and duplicate all the strings in values/
over in values-en/.

Lint already has a way to let you specify what your default language
is (e.g. tools:locale="es" on the root element in one of your strings.xml
files), but if it's not specified anywhere, default this to English.

Change-Id: I5b95d2d5f0a408f67adc5330d20e838c028745ba