Fix offset computation for attributes in command line XML parser
The position XML parser is a specialized DOM parser which keeps track
of offsets. This is used by lint and the manifest merger in order to
provide accurate error position information.

There is no SAX callback for individual attributes, so the way the
parser is able to compute accurate offset information for individual
attribute nodes is to keep the original XML source document around,
and when offsets are requested, it will search within the element
document range for the corresponding attribute declarations using
regexps to find the right tokens.

There were two bugs in this area, which manifested itself as the wrong
attributes being shown in lint error output messages (when lint is run
from the command line; in IDEs we use the native XML AST support which
does track offsets accurately) :

(1) When searching for an attribute declaration, if the target
    attribute does not have a namespace prefix, *and* the there is a
    different declaration of that attribute *with* a namespace prefix
    which is declared earlier, the parser would match the other
    (namespaced) attribute. In other words, if an element declares
    both "android:orientation" and "orientation", a node offset lookup
    for the "orientation" attribute would return the range for the
    "android:orientation" node.

    The fix for this bug is to make the search regexp explicitly make
    sure it does not have a prefix character before the name match.

(2) In some cases (with multiple attributes on a single line), it would
    include more than one attribute in the range. The reason for this
    was that the regexp would (after finding the opening quote) search
    for the closing quote. However, the regexp gladly matched not the
    next quote but the last one on the line.

    The fix here was to insert a reluctant qualifier into the regexp.

A number of unit tests were updated to reflect the now correct
attribute ranges. I spotted one error range being off in the
MissingDetectorTexts and went in and fix it, and I was really
surprised to see how many other places this had crept in without me
noticing!

Change-Id: I29a630d065b7c287092a75114d429fa0a00d8631