Fix FQCN to internal class name conversion.
When converting FQCN (class names in the form of foo.bar.Foo.Bar) to
internal names (of the form foo/bar/Foo$Bar), we assume that components
that start with upper case letters are class names.

However, we also use this method when we have an unambiguous FQCN of the
form foo.Bar.Foo$Bar. This changeset bypasses the heuristics when we
have an unambiguous case.

Bug: http://b.android.com/74140
Change-Id: Ic21aa2365949f9cf95ead1082c770750bd0527fc