Add build tools version as task input where needed.
This prevents incremental builds when switching between different
versions of build-tools.

Ideally, we'd like to declare the files themselves as task inputs. For
incremental builds, Gradle can't tell whether the removed file was a build
tool or a regular input file. Until this is fixed, it's rather tricky to
correctly handle build tools and input files in the same IncrementalTaskInputs
collection.

Bug: https://code.google.com/p/android/issues/detail?id=76201
Change-Id: I11407871cd11c4d5d3dfc4f89c8b7b953bf289e9