Fix pass at multi-dex for minSdkVersion 21+ app.
This requires build tools 21+ and is triggered by a multidex property
on build types or product flavors (build types have priority).

This does not handle pre-L compatibility yet, and the flag is ignored
if minSdkVersion is not 21+.

It only works for apps, so libraries with > 65K method will failed to be
pre-dexed still (this is the next step).

Change-Id: I5d621c140a43236e9cd85acfcf2fc93cb3358161