Fix for 76981 : injected package name should not create fake conflicts.
When comparing overlays and main manifest package names, we used the
injected value for the main package name creating an error. Instead,
use the original package name of the main manifest file to ensure it
is equal to the overlay one.

Also ensured that name expansion is using the original package name
rather than the injected value in overlays file.

Change-Id: Idf8637510574523d1619c060cfa1a095a3f96240