Fix resource shrinking for style references
If you have a theme reference which uses a dotted syntax
(e.g. MyOuter.Inner) that theme reference would not correctly
be resolved, meaning that some resources would not be marked
as reachable in the usage graph. This happened because resource
fields are renamed from A.B to A_B by aapt, and we need to
account for that in the resource URL lookup.

Change-Id: Ib8ccf35ac7871816fee1290860fcd2694c9f4793