78620: Wearable micro app is shrinked out by gradle plugin version 0.14.0
It looks like some runtime code loads an XML resource file
and within it finds the name of a raw resource to load (the APK).
However, that XML file is not using the normal AAPT resource syntax
(e.g. @raw/foo.apk), so the resource shrinking code did not treat
this as a valid resource reference.

Change-Id: I528447318fa892c37394cd564b3c9a66d9c70f84