Remove check for dependency on app project.
It turns out that we cannot just directly refuse an Android
app project as a dependency. The project could be publishing
other artifacts that we want to use.

Due to a limitation (or bug) in Gradle we cannot actually
check at the resolved dependency level, since additional
published artifacts get pull down with the default artifact
so the APK artifact shows up as well.

For now we'll just ignore the APK again until we can do better.

Added some (disabled) tests so that we can test this when
we figure out a better way to make it work.

Change-Id: I424c300dee3f9a9a70798cab2d68366a2753ceb8