Lint API check: Don't flag calling the overidden method
Lint flags all calls to methods that was introduced in an API level
higher than the minSdkVersion of your app.

However, if you are overriding a method that was introduced in a later
level, and your method just calls the overridden method from within
that body, don't flag that call: this method will probably only be
called in valid contexts.

Change-Id: I68d65d3085b7e17053dd3b79aed665ec62df0600