Fix problematic jar file leak.
URLConnections have caching enabled by default, and we're using them
to navigate to manifests inside jars. These jars end up kept
memory-mapped (and locked on Windows), in the cache, until someone
explictly calls close() or the garbage collector runs the finalizer.
Whether this happens fast enough in practice is unknown.

Meanwhile, if another process modifies the jar, the next attempt to read
crashes the JVM with a segfault.

In this CL, we effectively call close() on
sun.net.www.protocol.jar.JarURLConnection.JarURLInputStream, which takes
care of closing the JarFile underneath when caching is disabled.

Here's similar code in Gradle, in GUtil#loadProperties:
https://github.com/gradle/gradle/blob/master/subprojects/core/src/main/groovy/org/gradle/util/GUtil.java#L187

Bug: 17330223 (and maybe related external issues 76989 and 77602)
Change-Id: I571fb52566fcffd1558248c718747fd761316ea8