First version of the split-select tool integration.
Although the split-select tool is not fully functional yet, basic
wiring to get the devices characteristics through a single call to
aapt and feed the split-select tool with these characteristics.

This is WIP, for instance, we do not disambiguate yet if the APKs
are split enabled or not which would not require the split-select
tool.

The feature is turned off by default, set USE_SPLIT_SELECTOR to
true and use a 21 MR1 SDK.

Change-Id: Ic106f112a5fc41d14eb02f5c89434ab9d188363c