Fix ExecutorSingleton to be reused.
Starting with Gradle 2.3, the daemon can cache
classloaders to not have to reload plugin code
at every build. This did not go well with the
executor singleton that stops working after
shutdown is called at the end of the first build.

This makes it possible to get a new singleton after
the previous one was shutdown.

Change-Id: Ie504e73fa488a2ac082423b76d1b28d69d935fd9