Add two lint check for invalid BCP47 language/region codes
First, add a lint check which looks for invalid uses of
3 letter language or region codes where there is a corresponding
2 letter language or region code available.

Second, look for resource folders that are ignored by Gradle
(because the FolderConfiguration does not match), and flag this
as an ignored folder. Also look to see if it looks like an
incorrectly specified locale folder (e.g. using a 3 letter
language without using the b+ syntax) and if so make a more
specific suggestion.

Change-Id: I9ba31b1cfa9945d81c229d859e818d0bb1aac747