ddmlib: Fix getDensity() number format exception
Certain devices return a non-integer (e.g. 120i) for the system
property ro.sf.lcd_density. In such cases, we return -1 instead.
The default value when getDensity() method fails has also been
changed from 0 to -1. This is because a value of 0 could be assumed
to mean NO_DPI instead of an error condition.

Change-Id: I82d187b3244310a21aa273208def03a2d5200d1b