Added a task to jar the javac compiled task
provided an abstraction to the classes.jar provider.
wired that abstraction to the jack task that produces classes.zip
created an AndroidJarTask that takes the javac compiled classes
and bundle them up in a classes.jar.

This BinaryFileProviderTask abstraction will be used to declare
the necessary configuration dependency as well as retrieving the
bits for publishing the classes.jar as project artifact.

Change-Id: I2338967648dea0fe0105eeb32130035f2743bd3a