74584: Duplicated Icons under different names
First, allow some content-duplication under different names
if the name pattern is of a particular form (name_NNdp.png).

Second, if icons use that naming convention, check the actual
size implied by the density folder and the pixel size of the icon
and flag if it appears that the named dp size is wrong.

Change-Id: I9daa55048a344b227045a940e2f8452ac824965d