Don't skip any dependencies from testCompile.
We've had several bugs from people that can't find LogFactory.
commons-logging is currently skipped by DependencyChecker because Log
(but not LogFactory) is part of Android. This change removes Log from
the mockable android.jar, so libraries that detect they are on a JVM and
wich to use commons-logging can do so.

Bug: http://b.android.com/147673
Change-Id: Ia4a7cf703d87b9402ba7ea9c0f4a091fd544351d