Reuse Android Library objects in IDE model.
For very large projects with a lot of aar dependencies,
each model would create its own graph of dependencies with its
own instances. Because creating an AndroidLibrary instance includes
disk access (listing or searching for the .jar inside), this can
be slow if there are a lot of different projects re-using the
same aar. Using a cache significantly speeds this up (again, for
very large project).

Also move the model query action to be multi-threaded to speed up tests
(Studio will start doing the same soon).

While writing the test, I realized that it's not possible
to forward-create the internal configuration object used
for resolution, so it's not possible to create custom
resolution strategy. Fixed this by using maybeCreate()
instead.

Change-Id: I665c1746fc1659e35b4439bb6b52e910a218243b