Resource visibility lookup and lint check
This CL adds a mechanism to look up the visibility of resources (which
consults the public.txt files in AndroidLibraries). This is used by a
new lint check which flags any uses of private resources outside that
library.

There was an older private resources check, which looked for uses of
private framework resources. AAPT now enforces that check, so we don't
need to have lint do it anymore, so the detector has been repurposed
(and the issue severity adjusted) to look for library private
resources instead.

Change-Id: I3536f739ba8b04436907550001e39f8e08eadded