Unify implementations of SourcePosition (1st CL).
There are currently many different classes that represent a source file
and the location of some text in it, some of which 0-based, some
1-based.

This is the first CL in a series that aim to reduce this duplication and
the need for adapter classes.

The following immutable classes with Json serializers are defined:

SourceFile: contains a file and a description field for that file.
SourcePosition: a position range in a file.
SourceFilePosition: contains a SourceFile and a SourcePosition.
                    0-based, but the toString() method shows as 1-based.
Message: has a kind (ERROR, WARNING, ...), a descriptive message and a
         (non-empty) list of positions.

The manifest merger now uses these, other places to follow.

Change-Id: I262f64beee09d09320829bf07abf393bd82ac4e2