Delete data after doing resource merge.
While this would be collected at the end of the build,
we want to reduce memory usage during the build, which is
particularly important for builds that have multi-variants or
multi-modules.

This field actually hold the data set so it can be pretty big for
large project, so it's better to pro-actively allow it to be
collected.

Change-Id: I75c72fc6bbdcb852400e5cc4c3792a184dc1184a