Update ECJ from 4.4 to 4.4.2.
This includes a lot of bug fixes to the compiler (4.4 was released
in June 2014, 4.4.2 in March 2015. This might fix some crashes
reported in for example b.android.com/172268 and b/21667599.

This CL also tweaks the logging we perform when ECJ throws
exceptions: instead of logging the full exceptions, we only log
the first 20 files we encounter problems in, and we don't dump out
a full stack trace; instead, we only abbreviate the top 3-4
stack frames into a single line.

Change-Id: I42c9cc0b873e0e9687af66876c8346ec0bf7678b
(cherry picked from commit 0dc7c2ea2aebf440aac3b94becca79d423f52897)