ddmlib DeviceMonitor: Use a separate lock for selector registrations
The device monitor uses a selector to select between multiple
channels (one per device). Each channel needs to be first registered
with the selector. This registration can only happen when the selector
is not already waiting on the select() call.

In order to achieve this, registrations are performed while a special
lock is held, and the selector is not allowed to enter the select()
method unless the same lock is obtained. This prevents the main loop
from calling the select() method until the registration is complete.

This CL simply documents all this, and uses a new object as the special
lock rather than reusing some existing field.

Change-Id: Ia4b10ac5cc4a8f5ea80d335e060b179345a5bcf1