Avoids race condition that leads to NPE.
When a channel is closed it can still be in use. If this happens
then mChan is set to null and a null pointer exception happens as
synchronization is not done properly. This change is not trying
to fix the synchronization issues, but just avoid this particular
case by allowing the IOException of reading a closed socket to
happen anyway.

Change-Id: I7315e90f45b6dcec974058408048c5455cc568c0
(cherry picked from commit cdc4bc2089e20d6b76d5ad032bcdc344a40a04ec)