Change the new BlankActivity wizard to generate an App Bar.
The AppTheme for a module has been changed to include the 2 colors:
   colorprimary and colorPrimaryDark
this change makes the App Bar standout.

Added a common folder with recipes for an App Bar and a simple
RelativeLayout and values for both layouts.

The App Bar can be selected in the BlankActivity wizard (on by default).

Also: only add a menu and a "Hello world" TextView if this is a new
project (they will not be generated if this is a new activity).

Change-Id: I767a41b873442161129ae5c5104c635956ac1f07