Implement remaining encode and decode
This implements encode and decode for both map types and boxed types.
These are needed to decode the state blocks.
Encoding boxed numbers will always encode as unsigned because we can't
distinguish between signed and unsigned in java.
There is currently no support for encoding boxed slices, we don't need
it right now.

Change-Id: Ic05559d09a5aaaf536d1517715562cf6fa5767ad