Update lint StringFormatDetector's unit tests
StringFormatDetector had a number of unit tests which operated on
erroneous source files. This wasn't a problem earlier when lint
didn't have proper type resolution; as long as the AST was parseable
StringFormatDetector could operate on it (though it had to make
guesses about whether a method resolved to an actual string formatting
method.)  As part of fixing issue 183643 I plan to make it use the
resolved types to check, so in this CL I'm cleaning up the tests to
not have any compilation errors first. There are no semantic changes.

Change-Id: I90d2aaf3ed0116c04775677e751a3ea0cefcc962
(cherry picked from commit 10752f514a2adf83d046514c40ebfb6566da5915)