ddmlib: Provide a new API to add loggers
The old API was setting a single static variable (we have many
instances of such behavior in ddmlib), which meant that only a single
custom logger could be set. This CL deprecates that API and adds
a new one via which you can add multiple loggers if necessary.

This ofcourse is only necessary because in ddmlib, many APIs don't
provide detail about any errors that could occur, but instead log
those errors.

Change-Id: I953da122e14d9b7b82580516db9471cab406ef99