Move the FAB to the AppBar on a request from J.
I moved the FAB to the bottom right of the display on a request from
Chris Banes. However J feels like this example is specific for
scrolling behaviour and should show a different location for the
FAB. People can look at BlankActivy to see how to move it to the
bottom right if that is where they want it.

Change-Id: If6eb7b610b7dd257a3acbea5ddae502afc767b11