Adding duplicate Strings analysis on hprof captures.
This analyzer analyzes hprof captures for Strings that are duplicates
of each other. Duplicate Strings is a common memory waste in apps, and
this analyzer helps to find these wastes.

Also refactored the code to construct the Strings from the hprof dump
so that it's available to the analyzer as well.

Change-Id: I135d79a56151381fff812eab04ff8d34a1677bd7