Improves constructor support
Bytecode is now simulated to identify the right INVOKESPECIAL.
The invoke special might be a super class or the same class
if the constructor uses this(...) to delegate.

Removed "VisitorAdapter" that ignored part of the method.
Now a new method is generated that contains only the bits of
the bytecode that need to be translated.

Changed the SupportVisitor to use the same code to detect
the invokespecial, and use that to inject a redirection point
in the method, not just at the top of the method.

Change-Id: Idf95a69a2fc4b1db8cc8885b75b0fe2053d19657