Enable 3rd party transforms on libraries.
Transforms on libraries are limited to the scope
that is packaged with the aar (project, local_deps).

This has a significant impact on the way the aar is packaged
so there's a few things going on in there:

- This attempts to keep the separation of project code vs.
  local deps in order to speed up building aar in the sub-module
  use case. Once we can depend on folder across modules, this can
  be simplified.
- This now merges the java res always (before it was only if proguard
  was run). They get packaged with the main classes.jar, while
  the local deps only contain code. This is better as this puts
  the burden on conflict resolution on the author rather than
  the consumer.

This is achieve with a mix of regular Jar tasks, and
a custom transform creating separate jars for all the local deps,
and some copy/sync task to deal with the fact that we cannot
(at the moment?) control where the output of a transform is.

Also improves the checks on apk/aar for (doesNot)containClasses
methods to allow searching in a reduced scope. Makes it easier
to make sure a class is in the local jar rather than the main
classes.jar file.

(cherry-pick from 14533478233f4406a7d82669d3a411a365f00a00)

Change-Id: I112ba4006a478acb26569b79848a7750bdbbfa94