Support for package private instantiation/access.
After trying to intercept checkcast and/or force the checkcast in the
originating package, the VM still catches the class escaping the
package as an IllegalAccessError.

For now, reset packageprivate classes as public so they can be used
from the $override class directly. The vm already does that for inner
private classes which are in fact package private.

This mean we may have code that relies on getClass().getModifiers()
to return a package private value will fail.
several solutions :
   - look into loading the $override classes in the original class
   loader and therefore original package (renaming the $override
   every time).
   - intercept the getClass().getModifiers() call to return a saved
   copy of the modifiers
   - do not instant reload package private classes which use the
   Class.getModifiers() method.

 for now, reseting the flag will suffice, until we can look into
 reusing the original class loader possibility.

Change-Id: I4ceaebd6593063450a1853f4c3544c424722b911