updated transforms to latest API and added dependencies
dependencies were not added correctly to the transform which did not have
visibility on all the classes of the project. ASM needs to load some classes
to determine the common parent of 2 classes in order to compute the frames.

Set up a context class loader capable of loading classes from the project
set of classes and override the getCommonSuperClass method to use that
context call loader rather than the plugin class loader.

Change-Id: I8f17aac1d028712a3b2b16ade024eb200ca1d395