Improve time bound on dominator computation.
This CL replaces the previously used Cooper-Harvey-Kennedy dominator
computation algorithm with the much better worst-case time bound
Lengauer-Tarjan dominator algorithm. This solves the issue where a
large number of outstanding Finalizers in the Android GC chain could
cause the original dominator algorithm to run for hours on end.

This CL also adds a much finer-grained progress indicator for the
dominator computation to give a better indicator of the stages and
estimated progress until completion.

Furthermore, the reverse reference bookkeeping was mixed as part of
the visitor computation for the topological sort. This needed to be
refactored out to decouple the two systems. This allows the untangling
of the reference decoding from visitor computation, permitting other
systems to get the forward and backward references independently.

Tests have been included to verify the validity of the new algorithm
with respect to the old algorithm.

Change-Id: Ide84ef66f53719aafb428c6746412d96e5c4b2d6