Fix regression where package name was dropped
... and add test to prevent further regressions.

When a package name is unknown, we should show this:

"time pid-tid/? tag: message"

Instead, due to a refactoring, we're dropping the question mark:

"time pid-tid/ tag: message"

Previously, this logic was broken, but we handled it at a higher
level. Now, since we're relying more on the core classes to do all the
work, this bug surfaced.

Change-Id: I87bc05e2913f1d91fc7ae51d53438b7bfeccb5c6