Renamed "this" on overridden methods.
In order to be able to see the "this" variable, it needs
to be renamed to something different. Any choice of the form
"this$xyz" doesn't work either, in fact without any
BCI just naming a parameter "this$xyz" compiles and runs
fine, but the argument is not visible in the debugger.

Change-Id: Ic908b3dd20df37a393f5ff473487eaed5cce3962