Transform api revamp.
The previous API was very inflexible regarding
the output format. There is a need to have transforms
be able to output a mix of jars and folders.
In general a good practice is to output jars when
the inputs are jars and folders when the inputs are
folders.
It's particularly important to keep jars in case there
are case-sensitive conflicts inside the jars.

This new API solves this by making the output
allow any numbers of jars and/or folders, and tie them
to any combination of Types and Scopes. Because of this
(both the flexbility and the dynamicity), the API
also remove the need of having TransformOutputs matching
TransformInputs. There is a single TransformOutput
containing any thing in term of Format, Types, Scopes.

This also allows removing the different Transform
interfaces. Transforms don't have to specify their
types, they just create the outputs they want based
on what they do.
This prevents the setup check regarding scope (where
a single scope transform cannot follow a transform
that combines scopes), so this check is now at runtime.

In order to properly handle this new output type, a
new type of TransformStream is introduced. There are
now two types, one for things generated by a transform,
which is based on a single folder containing a specific
structure representing Format, Types, and Scopes; and
a stream simply containing files, all associated with
a specific Types/Scopes. This latter stream is used to
inject data generated by other tasks into the pipeline.

Change-Id: I1ce1e891b737a0e3a1990fcea6204cc39e83a197