Updated resource filtering.
If minSDK is defined, resources which could never be picked are
not included in the apk. For example, if minSDK is set to 21 and
there is a res/raw/file.txt and res/raw-v14/file.txt then the
former will be exluded from the apk because it will never be used.

Filtering only applies to resources that are single file, i.e.,
non-mergeable resources. Gains from filtering will significantly depend
on how many "useless" resources are there in supporting libraries.

For example, in iosched, 4 resources were not included in the final
apk (with some ridiculous savings of around 50b).

Some performance impact data in iosched:

Full build (gw --no-daemon clean assembleDebug):

+--ResourceMerger::mergeData............................  2,374,330,870 ns/i;     6 i
 |   +--ResourceMerger::filterAccept.....................   113,525,631 ns/i; 21560 i
 |   |   +--ResourceMerger::filterAccept/rebuildCache....    58,110,721 ns/i;     6 i

Merge data took 2.3s in the whole build with a total of 6 invocations.
We can see that filter accept increases mergeData time by 113ms (<5%).

Incremental build (gw --no-daemon):

ResourceMerger::mergeData was not invoked at all.

Change-Id: Iae80dd28982ee8824af6be67606142d5ca378e72
Bug: http://b.android.com/180609