Lint check to identify native code in unsafe locations
This lint check identifies native code located in the
application's res or assets directory (although the assets
directory currently isn't being scanned, at least when using
gradle). The preferred approach is to store the
native code in a directory that will result in it
being packaged into the compiled apk's lib directory.

The goal of this lint check is to encourage developers
to use shared libraries for their native code, and to
place those shared libraries such that they get stored
in the compiled apk's lib directory. Shared libraries
in the apk's lib directory are moved by the Android
Package Manager at install time to an application library
directory in /data/app-lib that applications themselves cannot
write to, decreasing the risk that the code will be
tampered with after install time.

This lint check may be able to help improve application
development practices so that execute permission on
application internal storage directories could be
removed in a future Android API level (similar to our
proposal here [1]).

[1] https://android-review.googlesource.com/#/c/169950/

Change-Id: I16b878616ef819418e5c523a3e8f8a6e91f7ba1b