Refactor to eliminate wasteful array copies.
getInstance() was performing an array copy under the hood. This often
meant simple calls such as getInstance().size() was needlessly copying
the whole array just to find the size. This CL removes the method in
favor of a Heap API more inline with the underlying data structure for
performance and memory consumption reasons.

Change-Id: I27b71e1cddb8deaeb13bed4587f0e027dc088fa8