Tweaks to the unsafe native code detector
Don't use an OtherFileScanner to find code in resource
and asset folders; use the project methods for iterating
these folders instead such that we can find for example
resource folders that have been renamed from the default
"res" name.

Some other misc cleanup and tests.

This also adds a lookup of the assets folders to the
lint project API.

Change-Id: I03a512eb12729cc970ef84ab420f029c83bb9e46