190812: Android Studio is unable to clean custom lint jar
Work around
http://bugs.java.com/bugdatabase/view_bug.do?bug_id=5041014

URLClassLoader, on Windows, locks the .jar file forever.  As of Java
7, there's a workaround: you can call close() when you're "done" with
the file. We'll do that here. However, the whole point of the
{@linkplain JarFileIssueRegistry} is that when lint is run over and
over again as the user is editing in the IDE and we're background
checking the code, we don't to keep loading the custom view classes
over and over again: we want to cache them. Therefore, just closing
the URLClassLoader right away isn't great either. However, it turns
out it's safe to close the URLClassLoader once you've loaded the
classes you need, since the URLClassLoader will continue to serve
those classes even after its close() methods has been called.

Therefore, if we can call close() on this URLClassLoader, we'll
proactively load all class files we find in the .jar file, then close
it.

(This CL also allows LintClient to provide the ClassLoader to use for
URLs. In Android Studio we'll use this hook to supply a non
URLClassLoader implementation, available in the IDE, which doesn't
lock the underlying resources. That prevents the IDE, often running on
Java 6, to lock the files.)

Change-Id: I03fd233d4fc2d0dd66fa01f742109123ea30613e