Setting permissions on downloaded files
Unfortunately in order to be able to read the permissions
attributes on files in zips you need to use the apache
non-streaming zip reader, since the permissions are stored
in the zip central directory rather than along with the file.
This means we need actual files to unzip rather than streams,
so the downloader interface is changed accordingly.

Change-Id: I1ea1e00055afa4229a58ea04a22323dda2a9a1fa