AndroidDebugBridge: Use CopyOnWriteSet instead of array list and locks
The AndroidDebugBridge class maintains 3 sets of listeners (bridge
changes, device changes and client changes). In all 3 cases, the
listeners were maintained in an ArrayList, but the accesses were
protected by a lock, and a copy of the list was being made every time an
event was being dispatched. This CL replaces all 3 instances with
CopyOnWriteArraySets which eliminates the need for the copying and the
lock in two of the cases.

Change-Id: I1f63d48d7dd70d564872d94330f2e9cfe23168dd