196925: Wrong NewApi error on FloatingActionButton#setBackgroundTintList()
Lint uses an API database to check which API level a given method was
introduced in.

When calling virtual methods, it also needs to look up the inheritance
chain to see if the method is defined on a framework class with a
given API requirement.

For example, ImageButton#setBackgroundTintList() was introduced in API
21.  If you have a custom view, called MyImageButton, which extends
ImageButton, and you then invoke
    myImageButton.setBackgroundTintList()
then this should generate an API warning because you're calling the
inherited method setBackgroundTintList().

However, the support library introduces a wrinkle in this approach,
because in many cases it extends an existing class, but backports
methods on it. In our example, FloatingActionButton is a subclass of
ImageButton, added to the design library, but its minSdk level is 7,
not 21. However, because of the above facility, as described in issue
196925 lint will flag calls to
FloatingActionButton#setBackgroundTintList() when minSdkVersion is
less than 21.

To fix this, lint now ships with its own API "overlay" for the support
library. It includes *only* those cases where we have a public support
library class that (1) extends a framework class, but is in a library
that has a lower minSdkVersion than the framework class, and (2)
provides methods that override framework methods but again with a
lower library minSdkVersion than the corresponding framework
method. This is currently the case for 32 classes (and 159 methods).

When generating the API database, it also pulls in these extra API
definitions and bakes them into its binary cache. For this reason, the
binary cache schema version is incremented from 7 to 8. The database
itself is generated by code in ApiLookupTest,
generateSupportLibraryFile().

Change-Id: I73ce93d241f01919ba80cd9b06e05e29c917e000