Support coldswap for API 19 through 23+
This pair of CLs change the instant run runtime library and IDE plugin
to support coldswap.

There are three different types of coldswap, based on the API level
of the target device:

(1) Marshmallow and up. For this, Gradle generates APK splits, which
    are pushed via adb install-multiple to the device. Also, the
    runtime library no longer applies the various class loader tricks and
    reflection hacks to set up overriding resources, which aren't necessary
    when using APK splits.

(2) Lollipop. For this, Gradle generates a number of dex shard files.
    The IDE plugin transfers these with fixed names (to ensure that
    subsequent changes to classes in the same shard will *overwrite*
    the previous shard.)

(3) KitKat. For this, Gradle generates a single restart dex file
    containing all the changed files since the last build.

We now also support "freeze swap" (also variously called dexswap),
which differs from coldswap in that we perform incremental builds even
when the app isn't running, and we push changed files to the device
via adb and then trigger starts via the activity manager.

There are a number of code cleanups to support all of this. We no
longer rely on the builder model to point to the restart/reload dex
files; instead, the build is fully driven off of the artifacts listed
in build-info.xml.

A few misc bug fixes too.

Change-Id: If2768841f351ec1f49c5a2bd6e4d801cff1eb499