Allow task execution when no code changes occured.
since we are delaying some artifacts construction when code changes
can be hotswapped, we need a way to manually kick some tasks when
the full artifacts are requested yet no code change occured.

Introduced a kicker task that will write a marker file containing
block or run commands. The downstream task/transform can use that
file as an input. The kicker task will change the content of the
file when the task must execute.

Change-Id: Ia07b17c6cd8d50328600abd170d59aaed8324623