Improved CachedFileContents change detection.
CachedFileContents relied on timestamps but this was not
completely reliable as in very fast changes (e.g., unit
tests) the timestamps may not change (ext3 has a time
resolution of 1s).

Now CachedFileContents uses several criteria to check
whether a file has changed: timestamp, size and CRC32
checksum. The CRC32 is fast and good enough for the purpose
we want.

Change-Id: Ibb7d979a123aed4b43bf599eea0024d101f949da