Profiler Plugin initial commit
The first pass of the profiler gradle plugin is a proof of concept
which, when applied to an Android project, looks for any class
called MainActivity and instruments its "onCreate" method to add a
log call to it.

In a followup CL, this logic will be updated to inject a server and
other support code that will be used to help profiling a user's
Android app.

Change-Id: I2b2f7d79a41ae34af4e229da09a18496991e3e4e