simplify slicing using the hashcode last digit for bucket ID
we used to have a slightly over engineered mechanism based on
package boundaries but there are some concerns it is not stable
or deterministic enough.

simplified by using the FQCN hashcode and extract the last digit
as the package id since the number of slices is set to 10.

Change-Id: I7485a2a5d87a38801973fe53fb5fb63aeb9b0ed8
(cherry picked from commit a0c7c5832f8313f2d0cdba6d88546d88fc24122d)