Add safeguard against invalid ECJ node offsets
For some reason, ECJ in some cases will provide an AST node whose
offsets and parenthesis count leads to impossible source ranges.
For example, in the integration test file
iosched/android/src/main/java/com/google/samples/apps/iosched/ui/MapFragment.java
there's this variable declaration:
   final View v = getView();
The variable node for "v" is listed as having source offsets 7866 to 7867, and
with a parenthesis count of 1 (this is the wrong part), which means we compute
it to have offsets 7867 to 7866 (an invalid range).

Change-Id: I6ebe5f303ac6351f4dd6f1be6a38b644e61d0bb4