Force creation of DEX file in case they are empty.
After switching to using package hashcode to select the slice
in which files will be copied to, it's more likely there will
be empty slices for small applications.

It then becomes difficult to maintain which slice should be
deleted from the device so it's better to always deliver 10
slices with one dummy class that will never get deleted.
Even when the last class from the slice is removed, a new
version of that slice will be delivered that will only
contain that guard class.

Change-Id: Ibc2ae2d671da16b8ce71352315ca76b69ee73f19