disable InstantRun hotswapping for constructors.
some open source libraries like Jacoco and Hugo are instrumenting the contructors in ways we are not expecting by declaring a local variable before the call to super is made.

we will re-enable it in the next release with a proper fix to our instrumentation code.

BUG=202347

Change-Id: Ib7d4faf0c2335f17b2ebe074511fe4286cc7f95d
(cherry picked from commit a9d1373d66feda48836000b154ba390adcc95602)