Ensure that the main process always wins the socket race
This CL makes the Bootstrap application, which is responsible for
starting the app server, look at the current process name and checks
whether it's the main process (which is the process which has the
application id as its package name). We can't just look for process
names that start with : since an app can define processes with other
package names too.

To be on the safe side, it will also start the server if there appears
to be only one process or if no process seems to have the package
name.

Change-Id: I5737aec1a6f8d3155a41c45b624b01cf8fb8f287