Update manifest merger with direct DOM access
When the manifest merger is run on the fly in the IDE,
we need access to the DOM. The manifest merger already has
this internally, but performs a full pretty print, exposes
that pretty printed XML to the IDE, which then turns around
and parses it again.

This CL adds an optional feature to turn off this pretty
printing intermediate step (not needed in the IDE), and
exposes a method to get direct access to the XmlDocument.

Also adds position-to-node lookup for the XML parser, and
underline tags to the HtmlBuilder.

Change-Id: Ie9e129762e97dc9043f1af20f735cf6f3c529711