Fix IDs generation in libraries.
We need the IDs in libraries to not be final. This
was the case for the library being generated but not for the
IDs of the transitive dependencies. This meant that if a library
imported the R class of its dependencies, then the non final
ID values would get inlined.

Change-Id: Ib6e761278d3626a8c56d7d3510aac5b646c5dddf