178582: Lint never returns
The DFS code in the ResourceCycleDetector was keeping track of
its path from the root and never analyzes into one of those nodes,
which means it avoids getting into an infinite cycle. However,
it wasn't keeping track of ALL visited nodes, which meant it could
visit a subgraph many times over. For large projects, this makes
a huge difference.

Change-Id: I97f4cc8128c02e2bcb1bf281d007f8f2ab87375c