Added virtual files to zip with extra field.
When filling in blank space in the zip file using extra
field, create virtual files that fill the blanks if they
are greater than 64k and cannot be covered by the extra
field. These virtual files are not indexed in the central
directory and should be invisible to any zip utility that
is standards-compliant.

Non-compliant tools, such as Oracle's jar tool when
expanding a zip file, will see files to the like of
_this_file_doesnt_exist :)

In general using these virtual files should be avoided as
it creates a weird user experience. However, they provide
an alternative to rebuilding the whole zip file.

Change-Id: Ia1072fb97dab8d381f1732fe947faf15b05a18fc