Adding Java grpc gradle and proto changes to profilers.
This adds the GRPC to profilers. The conflicts with protobuf2 are
resolved via jarjar'ing both the protobuf/grpc libs, as well as the
generated .java files from the .proto files. The results are then
jar'ed together into a fat jar, which will be consumed by Studio.

The proto changes are to fix issues intrinsic to Java grpc. Also,
the memory proto additions is for testing.

The actual inclusion and usage by Studio will be in a followup CL.

Change-Id: I41624141feca6e86fece730031e99d0ca5ba4a2e