Improve resource shrinking "is resource referenced from code?" heuristics
During resource shrinking, we look at the code to see whether it ever
ends up calling Resources.getIdentifier(). That method essentially
lets you do "reflection" on the resources: you can look up resources
by name, and that name can be computed on the fly. This defeats the
resource usage analysis, since we can end up removing resources that
are actually referenced from code.

To prevent this, we record the fact that the code is calling
getIdentifier, and we track all literal strings. Then, after looking
through the code, *IF* the code is making getIdentifier() calls, we
look through all the string literals found in the codebase and see
whether it looks like that string could be used to produce a resource
name.

For example, the code may be doing this:

    String name = String.format("my_avatar_%2d_thumb", avatar);
    getIdentifier( ... name ...)

The resource usage analyzer would come across the string literal
"my_avatar_%2d_thumb". It then looks through all available resources
and marks any resource that looks like it could be referenced from the
above; it does this by converting the above formatting string into a
regexp, and then matches that against all resource names.

Until this CL, we'd replace each formatting substitution string with
".*" -- match everything. However, we can do better than that. In the
above example, only *digits* can appear in the resource name, so we
shouldn't (for example) treat my_avatar_empty_thumb as a possible
match.

This CL changes the "formatting-string-to-regexp" conversion to treat
different formatting characters differently; %c will create a regexp
which only allows a single chracter; etc. This should help a lot with
overly broad formatting strings, like "%db" etc, e.g. a string
intended to print the number of bytes, suddenly matching ALL resource
names that end with "b".

Bug: 28618380

Change-Id: Idbc6af8d5ac38fec2481b24766f60c8b9d9fd8bd