Fix a linking issue regarding unit tests on Android
Previously the library link order was wrong resulting in a linkage
error when one of the static library dependency of the test depends on
functions from gnustl. This CL fixes the ordering issue with adiing the
library to the GTEST_LINK_LIBRARIES list and let the user specify the
link order for the libraries.

Change-Id: Ifabb0e1e8e3107c21d36d53f24defb8661cba52e