Better support for 16 & 32 bit images, fix Lab decoding
Gray and RGB images stored in 16 and 32 bit are now decoded as float
images to retain full precision (especially during color space
conversion).

This change also fixes the XYZ to sRGB step in the Lab color space
implementation. It was previously relying on the standard CIEXYZ
and sRGB color spaces but their XYZ to RGB step does not work with
our values (actually, rgbToXYZ(XYZtoRGB(x)) does not output the
input value as expected).

Change-Id: Ie174661c54d1a4bd9bbf9dd2c079ae9d2077b7fc