Improve the dependency handling for testdata
Previously we copied over testdata in configure time what meant that if
we added a new testdata filr or modified an existing one we had to do a
clean build to copy them over. This CL moves the copying to build time
what fixes the dependency handling issue but currently the copy happens
unconditionally what will have a performance overhead if we have a lot
of data (not the case at the moment).

Making the copy conditional without the explicit need for listing all
data files is fairly complicated so I plan to ignore it until copying
everything become a performance issue.

Change-Id: I375738a9d05c99bc20ccf342dc5a405dc64c203c