Cleanup instant run client <-> server communication
Issue 211989 shows a report where there was a communication error, yet
the IDE failed to show the communication error and reported that things
went well. This issue happened because when a communication error
happens, it was shown via a notification (callback on UserFeedback), but
the method itself did not return the error or throw an exception. This
is incorrect because: a) notifications can be disabled, and b) when
there is an error, the whole launch process needs to be terminated.

This CL fixes the issue by making sure that errors result in an
exception. The UserFeedback class was used in just 2 places and is
redundant, so it has been removed.

Addresses https://b.android.com/211989

Change-Id: If11d9b21e4ae53511a4b677412b55d3e3206f9d3