Fixing indentation in NewAndroidModule gradle template.
Workaround for https://code.google.com/p/android/issues/detail?id=213056

It seems http://ag/859276 changed the way the gradle files are
reformatted during the NPW. With that gradle files are getting
reformatted using TemplateUtils.reformatAndRearrange in
NewProjectWizardDynamic#runFinish, which is happening after the
GradleSyncState is updated. If the gralde files are updated due to
reformatting, the "Gralde files modified and project sync needed"
banner will be shown to the user.

While we investigate to find a proper fix, this change will be a
work around to avoid gradle file modifications during the reformat
operation.

Change-Id: I79f679c1096e9c82bbb7e1d8fba54f62a5e0e027