211348: Lint should not warn about appcompat compat attributes
We've special cased support for srcCompat, but generalize the
check such that we look for the origin of the attributes, and
if coming from appcompat, we accept them on non-appcompat tags.

Also clean up API to be very explicit about the kinds of resource
repository the lint rule wants: just the ones in the module,
or all modules, or modules + libraries.

Change-Id: I35d23770fd6bb54b67210bbed797addce3668d15
(cherry picked from commit 8ffb5f2ff6784cbd497829bf51c5798279cca1b6)