Tweak editDistance API to avoid computing unnecessary differences
The edit distance method is typically used to look for typo suggestions.
We generally only suggest a typo if two strings are very close (e.g.
an edit distance of 2 to 4 steps). If two strings are very different
(for example string length delta of 10 characters), we can already tell
that the edit distance will be much larger than the max delta we're
looking for: 3. So in that case, don't bother computing the edit
distance since the larger the delta, the more time and space it
requires to compute it.

To facilitate this, tweak the API such that the editDistance method
takes a max threshold, and add a "isEditableTo" method which lets you
pass in the threshold as well as compare with it, such that callers
don't need to repeat this value both in the edit distance call and in
the comparison check.

Change-Id: I7e3597be51ac9d6dbab7fc1c1e8cd723aaba1aae
(cherry picked from commit b82d0ca41368f840d29878407a151bae26899212)