Adding OfflineRepoCreator and extra repo sources
This change allows us to create a collection of zips and a
repository.xml that can be included in a build, and then referenced as
an SDK source to allow the included packages to be installed while
offline.
OfflineRepoCreator is the tool to create the repo.xml and zips in
prebuilts, and can be invoked using the updateOfflineRepo gradle task.
This change also adds an api to register new RepositorySourceProviders
that the RepoManager will query. The new source has to be defined at the
studio rather than base level, since the repo is included in studio
itself.

Change-Id: I7566ef5e06e4a37284ddcea92bb6df43f494afdd