194092: Lint does not add Java dependent projects to the compilation classpath
This CL changes the way initializes the graph of lint projects
from the corresponding Gradle project hierarchy when lint is running
inside the Gradle plugin.

Until now, it only used the builder model for the main app module
project; for the rest, it would just create library projects for
the corresponding AndroidLibrary and JavaLibrary dependencies.
But when these are local dependencies (e.g. local libraries, not
an AAR dependency) they don't actually provide references to the
real sources, only to the exploded AAR outputs -- which meant lint
wouldn't actually process the real source code.

To see why this is problematic, take a look at the regression test
included with this CL:
  integration-test/build/tests/LintLibraryModelTest/lintLibraryModel
That project contains lint errors in the Android and Java libraries,
but these are not found when you run :app:lintDebug.

This CL now looks up the right models for each library, and creates
wrapper projects for Java libraries (where there is no model builder.)

For now this is all behind a flag; to enable, run with
   lint.new-lib-model=true

Change-Id: I3285f88e79e348fd1c8177fb1033c9d08538bb3b
(cherry picked from commit a932297bf6e30f7eba463df4cbdbe347052596eb)