Add concurrency tests for FileCache.
The new ConcurrencyTester class is used to check whether or not a set of
runnables violate a concurrency contract (e.g., whether they run
concurrently when they are not allowed to, or vice versa). We currently
use it to write concurrency tests for FileCache.

Change-Id: I454efda0bdd792933d00b2b93152c715afb2eb5c
(cherry picked from commit fafd73470e16dcf398fabf19359505cf2c3d3602)