handle custom sourceSets for resources correctly
When custom sourceSets are added for resources which are actually children
of another sourceSet, the procedure to identify which sourceSet a file
belongs to inacurratly returned the first one instead of the longest one.

later on the file would be ignored as the MergeResource would disqualify the
file as not being part of the returned sourceSet.

This would effectively ignore all changes made to resources and continue
deploying the old resources.

For instance :
    sourceSets {
        main.res.srcDirs += 'src/main/res/layouts/shared'
    }
any file places into src/main/res/layouts/shared/layout would be wrongly
assigned to the src/main/res sourceSet instead of the one above.

BUG=219145

Change-Id: I31944a14cd4787bfef8cb77ee4f2c98c8cd71699
(cherry picked from commit 50e84ed6245a7c9f953e062ea98ae8d41a52aeba)