Fix broken dep on prepare library task.
The reverseMap fails to find the configuration associated
with a library because the library isn't immutable in the
case of a test dependency where the direct dependency is
a module that transitively brings dependency already
present in the app module (support lib for instance).
In this case the duplicated dependency gets marked as
skipped to avoid having the app package it twice and
this changes the key hashcode breaking the map.

The fix is to either change to a identity map or
to loop on the Map.Entry like this fix does (using
a identity multimap is a more complex change because
Guava doesn't provide one).

Ideally of course we'd have an immutable map entry
but it's not possible with the processing we have to
do.

Test: This needs new test coverage showing this particular scenario.
      Coming in a separate CL.

Bug: http://b.android.com/224325
Change-Id: I08619b89c11e489b417b71db751052a92eadbf32